Flytt overgangsstønad-spesifikke funksjoner ut i egen fil
